Move TextToPdfFragment to its own package
