+++ app/src/main/res/layout/fragment_text_to_pdf.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.TextToPdfFragment">
+    tools:context=".fragment.texttopdf.TextToPdfFragment">
 
     <android.support.v4.widget.NestedScrollView
         android:layout_width="match_parent"
